backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Demo backend for the Securing S3 Access tutorial"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.61",
    "fastapi>=0.135.1",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
